+++ MemoryGame/app/src/main/java/com/melihsurkmez/memorygame/GameForTwo.kt
@@ -339,7 +339,7 @@ class GameForTwo : AppCompatActivity() {
         var totalScore = puan.text.toString().toInt()
         var timeScore = sayac.text.toString().toInt()
 
-        var calculate = (card.score*2*card.home)*(timeScore/10)
+        var calculate = ((card.score*2*card.home)*(timeScore.toFloat()/10)).toInt()
         calculate = calculate + totalScore
 
 
@@ -363,11 +363,11 @@ class GameForTwo : AppCompatActivity() {
 
         if(card1.home == card2.home){
 
-            calculate = ((card1.score+card2.score)/card1.home)*(timeScore/10)
+            calculate = (((card1.score+card2.score)/card1.home)*(timeScore.toFloat()/10)).toInt()
             calculate = totalScore-calculate
 
         }else{
-            calculate = ((((card1.score+card2.score)/2)*card1.home*card2.home))*(timeScore/10)
+            calculate = (((((card1.score+card2.score)/2)*card1.home*card2.home))*(timeScore.toFloat()/10)).toInt()
             calculate = totalScore-calculate
 
         }
